refactor: expense -> allowance 이름 변경
