+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/DayDTO.kt
@@ -1,0 +1,3 @@
+package com.paykids.data.model.allowance
+
+class DayDTO : ArrayList<DayDTOItem>()
