+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/DayDTOItem.kt
@@ -1,0 +1,19 @@
+package com.paykids.data.model.allowance
+
+
+import com.google.gson.annotations.SerializedName
+
+data class DayDTOItem(
+    @SerializedName("allowanceType")
+    val allowanceType: String,
+    @SerializedName("amount")
+    val amount: Int,
+    @SerializedName("category")
+    val category: String,
+    @SerializedName("date")
+    val date: String,
+    @SerializedName("id")
+    val id: Int,
+    @SerializedName("memo")
+    val memo: String
+)
