+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/MonthAllCategoryDTO.kt
@@ -1,4 +1,4 @@
-package com.paykids.data.model.expense
+package com.paykids.data.model.allowance
 
 
 class MonthAllCategoryDTO : ArrayList<MonthAllCategoryDTOItem>()
