+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/MonthAllCategoryDTOItem.kt
@@ -1,9 +1,9 @@
-package com.paykids.data.model.expense
+package com.paykids.data.model.allowance
 
 
 import com.google.gson.annotations.SerializedName
 
-data class MonthMostCategoryDTO(
+data class MonthAllCategoryDTOItem(
     @SerializedName("allowanceType")
     val allowanceType: String,
     @SerializedName("amount")
