+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/MonthCategoryDTO.kt
@@ -1,4 +1,4 @@
-package com.paykids.data.model.expense
+package com.paykids.data.model.allowance
 
 
 class MonthCategoryDTO : ArrayList<MonthCategoryDTOItem>()
