+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/MonthDailyDTOItem.kt
@@ -1,17 +1,13 @@
-package com.paykids.data.model.expense
+package com.paykids.data.model.allowance
 
 
 import com.google.gson.annotations.SerializedName
 
-data class AddExpenseRequestDTO(
+data class MonthDailyDTOItem(
     @SerializedName("allowanceType")
     val allowanceType: String,
     @SerializedName("amount")
     val amount: Int,
-    @SerializedName("category")
-    val category: String,
     @SerializedName("date")
-    val date: String,
-    @SerializedName("memo")
-    val memo: String
+    val date: String
 )
