+++ PayKids/data/src/main/java/com/paykids/data/model/allowance/MonthMostCategoryDTO.kt
@@ -1,19 +1,15 @@
-package com.paykids.data.model.expense
+package com.paykids.data.model.allowance
 
 
 import com.google.gson.annotations.SerializedName
 
-data class MonthCategoryDTOItem(
+data class MonthMostCategoryDTO(
     @SerializedName("allowanceType")
     val allowanceType: String,
     @SerializedName("amount")
     val amount: Int,
     @SerializedName("category")
     val category: String,
-    @SerializedName("date")
-    val date: String,
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("memo")
-    val memo: String
+    @SerializedName("percent")
+    val percent: String
 )
