--- PayKids/data/src/main/java/com/paykids/data/model/expense/DayDTO.kt
@@ -1,3 +1,0 @@
-package com.paykids.data.model.expense
-
-class DayDTO : ArrayList<DayDTOItem>()
