--- PayKids/data/src/main/java/com/paykids/data/model/expense/MonthDailyDTOItem.kt
@@ -1,13 +1,0 @@
-package com.paykids.data.model.expense
-
-
-import com.google.gson.annotations.SerializedName
-
-data class MonthDailyDTOItem(
-    @SerializedName("allowanceType")
-    val allowanceType: String,
-    @SerializedName("amount")
-    val amount: Int,
-    @SerializedName("date")
-    val date: String
-)
