+++ PayKids/data/src/main/java/com/paykids/data/repository/ExpenseRepositoryImpl.kt
@@ -5,13 +5,13 @@ import com.paykids.data.mapper.toDailyInfoList
 import com.paykids.data.mapper.toDayInfoList
 import com.paykids.data.mapper.toMonthAllCategoryList
 import com.paykids.data.mapper.toMonthCategoryList
-import com.paykids.data.model.expense.AddExpenseRequestDTO
-import com.paykids.data.model.expense.UpdateExpenseRequestDTO
-import com.paykids.domain.model.expenseIncome.DayInfo
-import com.paykids.domain.model.expenseIncome.MonthAllCategoryInfo
-import com.paykids.domain.model.expenseIncome.MonthCategoryInfo
-import com.paykids.domain.model.expenseIncome.MonthDailyInfo
-import com.paykids.domain.model.expenseIncome.MonthMostCategory
+import com.paykids.data.model.allowance.AddExpenseRequestDTO
+import com.paykids.data.model.allowance.UpdateExpenseRequestDTO
+import com.paykids.domain.model.allowance.DayInfo
+import com.paykids.domain.model.allowance.MonthAllCategoryInfo
+import com.paykids.domain.model.allowance.MonthCategoryInfo
+import com.paykids.domain.model.allowance.MonthDailyInfo
+import com.paykids.domain.model.allowance.MonthMostCategory
 import com.paykids.domain.repository.ExpenseRepository
 import javax.inject.Inject
