+++ PayKids/data/src/main/java/com/paykids/data/service/ExpenseService.kt
@@ -1,13 +1,13 @@
 package com.paykids.data.service
 
 import com.paykids.data.model.BaseResponse
-import com.paykids.data.model.expense.DayDTO
-import com.paykids.data.model.expense.AddExpenseRequestDTO
-import com.paykids.data.model.expense.MonthAllCategoryDTO
-import com.paykids.data.model.expense.MonthCategoryDTO
-import com.paykids.data.model.expense.MonthDailyDTO
-import com.paykids.data.model.expense.MonthMostCategoryDTO
-import com.paykids.data.model.expense.UpdateExpenseRequestDTO
+import com.paykids.data.model.allowance.DayDTO
+import com.paykids.data.model.allowance.AddExpenseRequestDTO
+import com.paykids.data.model.allowance.MonthAllCategoryDTO
+import com.paykids.data.model.allowance.MonthCategoryDTO
+import com.paykids.data.model.allowance.MonthDailyDTO
+import com.paykids.data.model.allowance.MonthMostCategoryDTO
+import com.paykids.data.model.allowance.UpdateExpenseRequestDTO
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.DELETE
