+++ PayKids/domain/src/main/java/com/paykids/domain/model/allowance/MonthDailyInfo.kt
@@ -1,4 +1,4 @@
-package com.paykids.domain.model.expenseIncome
+package com.paykids.domain.model.allowance
 
 data class MonthDailyInfo(
     val date: String,
