+++ PayKids/domain/src/main/java/com/paykids/domain/model/allowance/MonthMostCategory.kt
@@ -1,4 +1,4 @@
-package com.paykids.domain.model.expenseIncome
+package com.paykids.domain.model.allowance
 
 data class MonthMostCategory(
     val category: String,
