+++ PayKids/domain/src/main/java/com/paykids/domain/repository/ExpenseRepository.kt
@@ -1,10 +1,10 @@
 package com.paykids.domain.repository
 
-import com.paykids.domain.model.expenseIncome.DayInfo
-import com.paykids.domain.model.expenseIncome.MonthAllCategoryInfo
-import com.paykids.domain.model.expenseIncome.MonthCategoryInfo
-import com.paykids.domain.model.expenseIncome.MonthDailyInfo
-import com.paykids.domain.model.expenseIncome.MonthMostCategory
+import com.paykids.domain.model.allowance.DayInfo
+import com.paykids.domain.model.allowance.MonthAllCategoryInfo
+import com.paykids.domain.model.allowance.MonthCategoryInfo
+import com.paykids.domain.model.allowance.MonthDailyInfo
+import com.paykids.domain.model.allowance.MonthMostCategory
 
 interface ExpenseRepository {
     suspend fun getMonthTotalExpense(
