+++ PayKids/domain/src/main/java/com/paykids/domain/usecase/expense/GetMonthAllCategoryUseCase.kt
@@ -1,6 +1,6 @@
 package com.paykids.domain.usecase.expense
 
-import com.paykids.domain.model.expenseIncome.MonthAllCategoryInfo
+import com.paykids.domain.model.allowance.MonthAllCategoryInfo
 import com.paykids.domain.repository.ExpenseRepository
 import javax.inject.Inject
