+++ PayKids/domain/src/main/java/com/paykids/domain/usecase/expense/GetMonthDailyExpenseUseCase.kt
@@ -1,6 +1,6 @@
 package com.paykids.domain.usecase.expense
 
-import com.paykids.domain.model.expenseIncome.MonthDailyInfo
+import com.paykids.domain.model.allowance.MonthDailyInfo
 import com.paykids.domain.repository.ExpenseRepository
 import javax.inject.Inject
